requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
google-genai==0.3.0
pandas==2.2.1
openpyxl==3.1.2
Pillow>=11.0.0
python-dotenv==1.0.1
rapidfuzz==3.6.1
